pyproject.toml
[project]
name = "feste"
version = "0.1.0"
description = "Reminds you."
readme = "README.md"
authors = [
    { name = "kuwifuwa" }
]
requires-python = ">=3.14"
dependencies = [
    "apscheduler>=3.11.0",
    "hikari>=2.4.1",
    "hikari-arc>=2.1.1",
    "pydantic>=2.12.3",
    "sqlalchemy>=2.0.44",
]

[project.scripts]
feste = "feste:main"
feste-init-db = "feste:init_db"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
hikari = { git = "https://github.com/mplatypus/hikari", rev = "feature/modals-v2" }
